[SQL] Clean up, comment aggregation transforms
