+++ sdks/java/extensions/sql/src/main/java/org/apache/beam/sdk/extensions/sql/impl/transform/agg/AggregationCombineFnAdapter.java
@@ -17,61 +17,57 @@
  */
 package org.apache.beam.sdk.extensions.sql.impl.transform.agg;
 
-import static java.util.stream.Collectors.toList;
-import static org.apache.beam.sdk.extensions.sql.impl.transform.BeamBuiltinAggregations.BUILTIN_AGGREGATOR_FACTORIES;
-import static org.apache.beam.sdk.schemas.Schema.toSchema;
-
-import java.io.Serializable;
-import java.util.List;
-import java.util.function.Function;
-import javax.annotation.Nullable;
 import org.apache.beam.sdk.coders.CannotProvideCoderException;
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.CoderRegistry;
-import org.apache.beam.sdk.coders.RowCoder;
 import org.apache.beam.sdk.extensions.sql.impl.UdafImpl;
+import org.apache.beam.sdk.extensions.sql.impl.transform.BeamBuiltinAggregations;
+import org.apache.beam.sdk.extensions.sql.impl.transform.agg.AggregationArgsAdapter.ArgsAdapter;
 import org.apache.beam.sdk.extensions.sql.impl.utils.CalciteUtils;
 import org.apache.beam.sdk.schemas.Schema;
-import org.apache.beam.sdk.schemas.SchemaCoder;
-import org.apache.beam.sdk.transforms.Combine;
+import org.apache.beam.sdk.transforms.Combine.CombineFn;
 import org.apache.beam.sdk.values.Row;
 import org.apache.calcite.rel.core.AggregateCall;
 import org.apache.calcite.sql.validate.SqlUserDefinedAggFunction;
 import org.apache.calcite.util.Pair;
 
 /**
- * Wrapper for aggregation function call. This is needed to avoid dealing with non-serializable
- * Calcite classes.
+ * Wrapper {@link CombineFn} for aggregation function call.
+ *
+ * <p>Delegates to the actual aggregation {@link CombineFn}, either built-in, or UDAF.
+ *
+ * <p>Actual aggregation {@link CombineFn CombineFns} expect their specific arguments, not the full
+ * input row. This class uses {@link ArgsAdapter arg adapters} to extract and map the call arguments
+ * to the {@link CombineFn CombineFn's} inputs.
  */
-public class AggregationCombineFnAdapter extends Combine.CombineFn<Row, Object, Object> {
+public class AggregationCombineFnAdapter extends CombineFn<Row, Object, Object> {
 
+  // Field for a function call
   private Schema.Field field;
 
-  private List<Integer> args;
+  // Actual aggregation CombineFn
+  private CombineFn combineFn;
 
-  private Combine.CombineFn combineFn;
-
-  private Schema sourceSchema;
-
+  // Adapter to convert input Row to CombineFn's arguments
   private ArgsAdapter argsAdapter;
 
+  /** {@link Schema.Field} with this function call. */
   public Schema.Field field() {
     return field;
   }
 
-  public AggregationCombineFnAdapter(
-      Schema.Field field,
-      List<Integer> args,
-      Combine.CombineFn combineFn,
-      Schema sourceSchema,
-      ArgsAdapter argsAdapter) {
+  private AggregationCombineFnAdapter(
+      Schema.Field field, CombineFn combineFn, ArgsAdapter argsAdapter) {
     this.field = field;
-    this.args = args;
     this.combineFn = combineFn;
-    this.sourceSchema = sourceSchema;
     this.argsAdapter = argsAdapter;
   }
 
+  /**
+   * Creates an instance of {@link AggregationCombineFnAdapter}.
+   *
+   * @param callWithAlias Calcite's output, represents a function call paired with its field alias
+   */
   public static AggregationCombineFnAdapter of(
       Pair<AggregateCall, String> callWithAlias, Schema inputSchema) {
     AggregateCall call = callWithAlias.getKey();
@@ -80,32 +76,21 @@ public class AggregationCombineFnAdapter extends Combine.CombineFn<Row, Object, 
 
     return new AggregationCombineFnAdapter(
         field,
-        call.getArgList(),
         createCombineFn(call, field, functionName),
-        inputSchema,
-        createArgsAdapter(call.getArgList(), inputSchema));
+        AggregationArgsAdapter.of(call.getArgList(), inputSchema));
   }
 
-  private static ArgsAdapter createArgsAdapter(List<Integer> argList, Schema inputSchema) {
-    if (argList.size() == 0) {
-      return new ZeroArgsAdapter(inputSchema);
-    } else if (argList.size() == 1) {
-      return new SingleArgAdapter(inputSchema, argList);
-    } else {
-      return new MultiArgsAdapter(inputSchema, argList);
-    }
-  }
-
-  private static Combine.CombineFn<?, ?, ?> createCombineFn(
+  /** Creates either a UDAF or a built-in {@link CombineFn}. */
+  private static CombineFn<?, ?, ?> createCombineFn(
       AggregateCall call, Schema.Field field, String functionName) {
     if (call.getAggregation() instanceof SqlUserDefinedAggFunction) {
       return getUdafCombineFn(call);
     }
 
-    return createBuiltinCombineFn(functionName, field.getType().getTypeName());
+    return BeamBuiltinAggregations.create(functionName, field.getType().getTypeName());
   }
 
-  private static Combine.CombineFn<?, ?, ?> getUdafCombineFn(AggregateCall call) {
+  private static CombineFn<?, ?, ?> getUdafCombineFn(AggregateCall call) {
     try {
       return ((UdafImpl) ((SqlUserDefinedAggFunction) call.getAggregation()).function)
           .getCombineFn();
@@ -114,25 +99,18 @@ public class AggregationCombineFnAdapter extends Combine.CombineFn<Row, Object, 
     }
   }
 
-  private static Combine.CombineFn<?, ?, ?> createBuiltinCombineFn(
-      String functionName, Schema.TypeName fieldTypeName) {
-
-    Function<Schema.TypeName, Combine.CombineFn<?, ?, ?>> aggregatorFactory =
-        BUILTIN_AGGREGATOR_FACTORIES.get(functionName);
-
-    if (aggregatorFactory != null) {
-      return aggregatorFactory.apply(fieldTypeName);
-    }
-
-    throw new UnsupportedOperationException(
-        String.format("Aggregator [%s] is not supported", functionName));
-  }
-
   @Override
   public Object createAccumulator() {
     return combineFn.createAccumulator();
   }
 
+  /**
+   * Calls the args adapter to extract the fields from the input row and pass them into the actual
+   * {@link CombineFn}. E.g. input of a MAX(f) is not a full row, but just a number.
+   *
+   * <p>If argument is null, skip it and return the original accumulator. This is what SQL
+   * aggregations are supposed to do.
+   */
   @Override
   public Object addInput(Object accumulator, Row input) {
     Object argsValues = argsAdapter.getArgsValues(input);
@@ -149,98 +127,15 @@ public class AggregationCombineFnAdapter extends Combine.CombineFn<Row, Object, 
     return combineFn.extractOutput(accumulator);
   }
 
+  /**
+   * {@link CombineFn#getAccumulatorCoder} is supposed to use input {@link Coder coder} to infer the
+   * {@link Coder coder} for the accumulator. Here we call the args adapter to get the input coder
+   * for the delegate {@link CombineFn}.
+   */
   @Override
   public Coder<Object> getAccumulatorCoder(CoderRegistry registry, Coder<Row> inputCoder)
       throws CannotProvideCoderException {
 
     return combineFn.getAccumulatorCoder(registry, argsAdapter.getArgsValuesCoder());
-  }
-
-  static class ZeroArgsAdapter implements ArgsAdapter {
-    Schema sourceSchema;
-
-    public ZeroArgsAdapter(Schema sourceSchema) {
-      this.sourceSchema = sourceSchema;
-    }
-
-    @Nullable
-    @Override
-    public Object getArgsValues(Row input) {
-      return input.getValue(0);
-    }
-
-    @Override
-    public Coder getArgsValuesCoder() {
-      return RowCoder.coderForFieldType(sourceSchema.getField(0).getType());
-    }
-  }
-
-  static class SingleArgAdapter implements ArgsAdapter {
-    Schema sourceSchema;
-    List<Integer> argsIndicies;
-
-    public SingleArgAdapter(Schema sourceSchema, List<Integer> argsIndicies) {
-      this.sourceSchema = sourceSchema;
-      this.argsIndicies = argsIndicies;
-    }
-
-    @Nullable
-    @Override
-    public Object getArgsValues(Row input) {
-      return input.getValue(argsIndicies.get(0));
-    }
-
-    @Override
-    public Coder getArgsValuesCoder() {
-      int fieldIndex = argsIndicies.get(0);
-      return RowCoder.coderForFieldType(sourceSchema.getField(fieldIndex).getType());
-    }
-  }
-
-  static class MultiArgsAdapter implements ArgsAdapter {
-    Schema sourceSchema;
-    List<Integer> argsIndicies;
-
-    public MultiArgsAdapter(Schema sourceSchema, List<Integer> argsIndicies) {
-      this.sourceSchema = sourceSchema;
-      this.argsIndicies = argsIndicies;
-    }
-
-    @Nullable
-    @Override
-    public Object getArgsValues(Row input) {
-
-      List<Object> argsValues = argsIndicies.stream().map(input::getValue).collect(toList());
-
-      if (argsValues.contains(null)) {
-        return null;
-      }
-
-      Schema argsSchema =
-          argsIndicies
-              .stream()
-              .map(fieldIndex -> input.getSchema().getField(fieldIndex))
-              .collect(toSchema());
-
-      return Row.withSchema(argsSchema).addValues(argsValues).build();
-    }
-
-    @Override
-    public Coder getArgsValuesCoder() {
-      Schema argsSchema =
-          argsIndicies
-              .stream()
-              .map(fieldIndex -> sourceSchema.getField(fieldIndex))
-              .collect(toSchema());
-
-      return SchemaCoder.of(argsSchema);
-    }
-  }
-
-  interface ArgsAdapter extends Serializable {
-    @Nullable
-    Object getArgsValues(Row input);
-
-    Coder getArgsValuesCoder();
   }
 }
